+++ apidaora/exceptions.py
@@ -1,5 +1,5 @@
 from dataclasses import dataclass
-from typing import Any, Dict, Optional, Sequence, Type
+from typing import Any, Dict, Optional, Sequence
 
 from .header import _Header
 
@@ -28,7 +28,7 @@ class InvalidReturnError(APIDaoraError):
 @dataclass
 class BadRequestError(APIDaoraError):
     name: str
-    info: Dict[str, Type[Any]]
+    info: Dict[str, Any]
     headers: Optional[Sequence[_Header]] = None
 
     def __str__(self) -> str:
@@ -37,3 +37,11 @@ class BadRequestError(APIDaoraError):
     @property
     def dict(self) -> Dict[str, Any]:
         return {'name': self.name, 'info': self.info}
+
+
+class InvalidTasksRepositoryError(APIDaoraError):
+    ...
+
+
+class InvalidRouteArgumentsError(APIDaoraError):
+    ...