+++ apidaora/route/decorator.py
@@ -1,8 +1,9 @@
 import functools
-from typing import Any, Callable
+from typing import Any, Callable, Union
 
-from ..asgi.base import ASGICallable
-from ..exceptions import MethodNotFoundError
+from ..asgi.router import Controller
+from ..controllers.background_task import BackgroundTask, make_background_task
+from ..exceptions import InvalidRouteArgumentsError, MethodNotFoundError
 from ..method import MethodType
 from .factory import make_route
 
@@ -12,20 +13,41 @@ class _RouteDecorator:
         if attr_name == '__name__':
             return type(self).__name__
 
-        method = attr_name.upper()
+        if attr_name == 'background':
+            brackground = True
 
-        if method not in MethodType.__members__:
-            raise MethodNotFoundError(attr_name)
+        else:
+            brackground = False
+            method = attr_name.upper()
+
+            if method not in MethodType.__members__:
+                raise MethodNotFoundError(attr_name)
 
         def decorator(
-            path_pattern: str,
-        ) -> Callable[[Callable[..., Any]], ASGICallable]:
+            path_pattern: str, **kwargs: Any
+        ) -> Callable[[Callable[..., Any]], Controller]:
+            if len(kwargs) > 0 and tuple(kwargs.keys()) != (
+                'tasks_repository',
+            ):
+                raise InvalidRouteArgumentsError(kwargs)
+
             @functools.wraps(make_route)
-            def wrapper(controller: Callable[..., Any]) -> ASGICallable:
-                route = make_route(
-                    path_pattern, MethodType[method], controller
-                )
-                return route.controller  # type: ignore
+            def wrapper(
+                controller: Callable[..., Any]
+            ) -> Union[Controller, BackgroundTask]:
+                if brackground:
+                    tasks_repository = kwargs.get('tasks_repository')
+                    return make_background_task(
+                        controller,
+                        path_pattern,
+                        tasks_repository=tasks_repository,
+                    )
+
+                else:
+                    route = make_route(
+                        path_pattern, MethodType[method], controller
+                    )
+                    return route.controller
 
             return wrapper